Add files via upload
Small revisions in punctuation and clarification of language without a change in content. Fixed issue with the header. Yet to fix the issue in the footer where copyright date reverts to xxxx-xxx. Changes made with Carl's agreement. Changed name of file to S-9.1.1.3 21 MTC... to be consiostent with the title.